pyproject.toml
[tool.poetry]
name = "djangoninja-learn"
version = "0.1.0"
description = ""
authors = ["Foshati <amirrezafoshati@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
django = "^5.0.7"
django-ninja = "^1.2.1"
django-allauth = "^0.63.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
